+++ rasa_core/actions/action.py
@@ -20,6 +20,8 @@ ACTION_LISTEN_NAME = "action_listen"
 
 ACTION_RESTART_NAME = "action_restart"
 
+ACTION_DEFAULT_FALLBACK_NAME = "action_default_fallback"
+
 
 class Action(object):
     """Next action to be taken in response to a dialogue state."""
@@ -103,3 +105,19 @@ class ActionRestart(Action):
 
         dispatcher.utter_template("utter_restart", tracker, silent_fail=True)
         return [Restarted()]
+
+
+class ActionDefaultFallback(Action):
+    """Executes the fallback action and goes back to the previous state
+    of the dialogue"""
+
+    def name(self):
+        return ACTION_DEFAULT_FALLBACK_NAME
+
+    def run(self, dispatcher, tracker, domain):
+        from rasa_core.events import UserUtteranceReverted
+
+        if domain.random_template_for("utter_default") is not None:
+            dispatcher.utter_template("utter_default")
+
+        return [UserUtteranceReverted()]